# IOData improvements and faster ghc-pkg dump parsing