--- a/Cabal/Distribution/Simple/Program/Run.hs
+++ b/Cabal/Distribution/Simple/Program/Run.hs
@@ -1,5 +1,6 @@
 {-# LANGUAGE FlexibleContexts #-}
-{-# LANGUAGE RankNTypes #-}
+{-# LANGUAGE GADTs            #-}
+{-# LANGUAGE RankNTypes       #-}
 
 -----------------------------------------------------------------------------
 -- |
@@ -22,24 +23,26 @@ module Distribution.Simple.Program.Run (
 
     runProgramInvocation,
     getProgramInvocationOutput,
+    getProgramInvocationLBS,
     getProgramInvocationOutputAndErrors,
 
     getEffectiveEnvironment,
   ) where
 
-import Prelude ()
 import Distribution.Compat.Prelude
+import Prelude ()
 
+import Distribution.Compat.Environment
 import Distribution.Simple.Program.Types
 import Distribution.Simple.Utils
-import Distribution.Verbosity
-import Distribution.Compat.Environment
 import Distribution.Utils.Generic
+import Distribution.Verbosity
 
-import qualified Data.Map as Map
+import System.Exit     (ExitCode (..), exitWith)
 import System.FilePath
-import System.Exit
-         ( ExitCode(..), exitWith )
+
+import qualified Data.ByteString.Lazy as LBS
+import qualified Data.Map             as Map
 
 -- | Represents a specific invocation of a specific program.
 --
@@ -55,17 +58,18 @@ data ProgramInvocation = ProgramInvocation {
        -- Extra paths to add to PATH
        progInvokePathEnv :: [FilePath],
        progInvokeCwd   :: Maybe FilePath,
-       progInvokeInput :: Maybe String,
-       progInvokeInputEncoding  :: IOEncoding,
+       progInvokeInput :: Maybe IOData,
+       progInvokeInputEncoding  :: IOEncoding, -- ^ TODO: remove this, make user decide when constructing 'progInvokeInput'.
        progInvokeOutputEncoding :: IOEncoding
      }
 
 data IOEncoding = IOEncodingText   -- locale mode text
                 | IOEncodingUTF8   -- always utf8
 
-encodeToIOData :: IOEncoding -> String -> IOData
-encodeToIOData IOEncodingText = IODataText
-encodeToIOData IOEncodingUTF8 = IODataBinary . toUTF8LBS
+encodeToIOData :: IOEncoding -> IOData -> IOData
+encodeToIOData _              iod@(IODataBinary _) = iod
+encodeToIOData IOEncodingText iod@(IODataText _)   = iod
+encodeToIOData IOEncodingUTF8 (IODataText str)     = IODataBinary (toUTF8LBS str)
 
 emptyProgramInvocation :: ProgramInvocation
 emptyProgramInvocation =
@@ -156,32 +160,41 @@ getProgramInvocationOutput verbosity inv = do
       die' verbosity $ "'" ++ progInvokePath inv ++ "' exited with an error:\n" ++ errors
     return output
 
+getProgramInvocationLBS :: Verbosity -> ProgramInvocation -> IO LBS.ByteString
+getProgramInvocationLBS verbosity inv = do
+    (output, errors, exitCode) <- getProgramInvocationIODataAndErrors verbosity inv IODataModeBinary
+    when (exitCode /= ExitSuccess) $
+      die' verbosity $ "'" ++ progInvokePath inv ++ "' exited with an error:\n" ++ errors
+    return output
 
 getProgramInvocationOutputAndErrors :: Verbosity -> ProgramInvocation
                                     -> IO (String, String, ExitCode)
-getProgramInvocationOutputAndErrors verbosity
-  ProgramInvocation {
-    progInvokePath  = path,
-    progInvokeArgs  = args,
-    progInvokeEnv   = envOverrides,
-    progInvokePathEnv = extraPath,
-    progInvokeCwd   = mcwd,
-    progInvokeInput = minputStr,
-    progInvokeOutputEncoding = encoding
-  } = do
-    let mode = case encoding of IOEncodingUTF8 -> IODataModeBinary
-                                IOEncodingText -> IODataModeText
+getProgramInvocationOutputAndErrors verbosity inv = case progInvokeOutputEncoding inv of
+    IOEncodingText -> do
+        (output, errors, exitCode) <- getProgramInvocationIODataAndErrors verbosity inv IODataModeText
+        return (output, errors, exitCode)
+    IOEncodingUTF8 -> do
+        (output', errors, exitCode) <- getProgramInvocationIODataAndErrors verbosity inv IODataModeBinary
+        return (normaliseLineEndings (fromUTF8LBS output'), errors, exitCode)
 
-        decode (IODataBinary b) = normaliseLineEndings (fromUTF8LBS b)
-        decode (IODataText   s) = s
-
+getProgramInvocationIODataAndErrors
+    :: KnownIODataMode mode => Verbosity -> ProgramInvocation -> IODataMode mode
+    -> IO (mode, String, ExitCode)
+getProgramInvocationIODataAndErrors
+  verbosity
+  ProgramInvocation
+    { progInvokePath          = path
+    , progInvokeArgs          = args
+    , progInvokeEnv           = envOverrides
+    , progInvokePathEnv       = extraPath
+    , progInvokeCwd           = mcwd
+    , progInvokeInput         = minputStr
+    , progInvokeInputEncoding = encoding
+    }
+  mode = do
     pathOverride <- getExtraPathEnv envOverrides extraPath
     menv <- getEffectiveEnvironment (envOverrides ++ pathOverride)
-    (output, errors, exitCode) <- rawSystemStdInOut verbosity
-                                    path args
-                                    mcwd menv
-                                    input mode
-    return (decode output, errors, exitCode)
+    rawSystemStdInOut verbosity path args mcwd menv input mode
   where
     input = encodeToIOData encoding <$> minputStr