--- a/Cabal/Distribution/Simple/Utils.hs
+++ b/Cabal/Distribution/Simple/Utils.hs
@@ -1,4 +1,5 @@
 {-# LANGUAGE CPP #-}
+{-# LANGUAGE GADTs #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE RankNTypes #-}
@@ -60,7 +61,8 @@ module Distribution.Simple.Utils (
         -- "Distribution.Utils.IOData" for convience as they're
         -- exposed in the API of 'rawSystemStdInOut'
         IOData(..),
-        IODataMode(..),
+        KnownIODataMode (..),
+        IODataMode (..),
 
         -- * copying files
         createDirectoryIfMissingVerbose,
@@ -168,10 +170,9 @@ module Distribution.Simple.Utils (
 
 import Prelude ()
 import Distribution.Compat.Prelude
-import Control.Exception (SomeException)
 
 import Distribution.Utils.Generic
-import Distribution.Utils.IOData (IOData(..), IODataMode(..))
+import Distribution.Utils.IOData (IOData(..), IODataMode (..), KnownIODataMode (..))
 import qualified Distribution.Utils.IOData as IOData
 import Distribution.ModuleName as ModuleName
 import Distribution.System
@@ -226,6 +227,7 @@ import System.IO.Unsafe
     ( unsafeInterleaveIO )
 import qualified Control.Exception as Exception
 
+import Foreign.C.Error (Errno (..), ePIPE)
 import Data.Time.Clock.POSIX (getPOSIXTime, POSIXTime)
 import Control.Exception (IOException, evaluate, throwIO, fromException)
 import Numeric (showFFloat)
@@ -234,6 +236,8 @@ import qualified System.Process as Process
 import System.Process
          ( ProcessHandle, createProcess, rawSystem, runInteractiveProcess
          , showCommandForUser, waitForProcess)
+
+import qualified GHC.IO.Exception as GHC
 
 import qualified Text.PrettyPrint as Disp
 
@@ -792,11 +796,10 @@ createProcessWithEnv verbosity path args mcwd menv inp out err = withFrozenCallS
 --
 -- The output is assumed to be text in the locale encoding.
 --
-rawSystemStdout :: Verbosity -> FilePath -> [String] -> IO String
+rawSystemStdout :: forall mode. KnownIODataMode mode => Verbosity -> FilePath -> [String] -> IO mode 
 rawSystemStdout verbosity path args = withFrozenCallStack $ do
-  (IODataText output, errors, exitCode) <- rawSystemStdInOut verbosity path args
-                                                  Nothing Nothing
-                                                  Nothing IODataModeText
+  (output, errors, exitCode) <- rawSystemStdInOut verbosity path args
+    Nothing Nothing Nothing (IOData.iodataMode :: IODataMode mode)
   when (exitCode /= ExitSuccess) $
     die' verbosity errors
   return output
@@ -805,15 +808,16 @@ rawSystemStdout verbosity path args = withFrozenCallStack $ do
 -- also supply some input. Also provides control over whether the binary/text
 -- mode of the input and output.
 --
-rawSystemStdInOut :: Verbosity
+rawSystemStdInOut :: KnownIODataMode mode
+                  => Verbosity
                   -> FilePath                 -- ^ Program location
                   -> [String]                 -- ^ Arguments
                   -> Maybe FilePath           -- ^ New working dir or inherit
                   -> Maybe [(String, String)] -- ^ New environment or inherit
                   -> Maybe IOData             -- ^ input text and binary mode
-                  -> IODataMode               -- ^ output in binary mode
-                  -> IO (IOData, String, ExitCode) -- ^ output, errors, exit
-rawSystemStdInOut verbosity path args mcwd menv input outputMode = withFrozenCallStack $ do
+                  -> IODataMode mode          -- ^ iodata mode, acts as proxy
+                  -> IO (mode, String, ExitCode) -- ^ output, errors, exit
+rawSystemStdInOut verbosity path args mcwd menv input _ = withFrozenCallStack $ do
   printRawCommandAndArgs verbosity path args
 
   Exception.bracket
@@ -828,16 +832,11 @@ rawSystemStdInOut verbosity path args mcwd menv input outputMode = withFrozenCal
       -- fork off a couple threads to pull on the stderr and stdout
       -- so if the process writes to stderr we do not block.
 
-      withAsyncNF (hGetContents errh) $ \errA -> withAsyncNF (IOData.hGetContents outh outputMode) $ \outA -> do
+      withAsyncNF (hGetContents errh) $ \errA -> withAsyncNF (IOData.hGetIODataContents outh) $ \outA -> do
         -- push all the input, if any
-        case input of
-          Nothing        -> return ()
-          Just inputData -> do
-            -- input mode depends on what the caller wants
-            -- todo: ignoreSigPipe
-            IOData.hPutContents inh inputData
-            --TODO: this probably fails if the process refuses to consume
-            -- or if it closes stdin (eg if it exits)
+        ignoreSigPipe $ case input of
+          Nothing        -> hClose inh
+          Just inputData -> IOData.hPutContents inh inputData
 
         -- wait for both to finish
         mberr1 <- waitCatch outA
@@ -855,8 +854,8 @@ rawSystemStdInOut verbosity path args mcwd menv input outputMode = withFrozenCal
                             " with error message:\n" ++ err
                          ++ case input of
                               Nothing       -> ""
-                              Just d | IOData.null d -> ""
-                              Just (IODataText inp) -> "\nstdin input:\n" ++ inp
+                              Just d | IOData.null d  -> ""
+                              Just (IODataText inp)   -> "\nstdin input:\n" ++ inp
                               Just (IODataBinary inp) -> "\nstdin input (binary):\n" ++ show inp
 
         -- Check if we we hit an exception while consuming the output
@@ -865,11 +864,17 @@ rawSystemStdInOut verbosity path args mcwd menv input outputMode = withFrozenCal
 
         return (out, err, exitcode)
   where
-    reportOutputIOError :: Either SomeException a -> NoCallStackIO a
+    reportOutputIOError :: Either Exception.SomeException a -> NoCallStackIO a
     reportOutputIOError (Right x) = return x
     reportOutputIOError (Left exc) = case fromException exc of
         Just ioe -> throwIO (ioeSetFileName ioe ("output of " ++ path))
         Nothing  -> throwIO exc
+
+    ignoreSigPipe :: NoCallStackIO () -> NoCallStackIO ()
+    ignoreSigPipe = Exception.handle $ \e -> case e of
+        GHC.IOError { GHC.ioe_type  = GHC.ResourceVanished, GHC.ioe_errno = Just ioe }
+            | Errno ioe == ePIPE -> return ()
+        _ -> throwIO e
 
 -- | Look for a program and try to find it's version number. It can accept
 -- either an absolute path or the name of a program binary, in which case we