… Cabal/tests/ParserTests.hs
… Cabal/tests/UnitTests/Distribution/Simple/Utils.hs
… cabal-install/Distribution/Client/Install.hs
… cabal-install/Distribution/Client/ProjectBuilding.hs