Based on #6364 

Parsing the `ghc-dump` is now two times faster and uses less memory.
For big cabal nix-stores this speedup is noticeable:

```haskell
{- ghc-pkg --simple-output

%  time ghc-pkg-8.6.5 --package-db=/cabal/store/ghc-8.6.5/package.db list --simple-output | wc 
      1    4473   82025
ghc-pkg-8.6.5 --package-db=/cabal/store/ghc-8.6.5/package.db list --simple-output  2,40s user 0,06s system 99% cpu 2,457 total
wc  0,00s user 0,00s system 0% cpu 2,457 total

-}
{- This script BEFORE THE CHANGE

% time runghc GhcPkgScript.hs ghc-pkg-8.6.5 /cabal/store/ghc-8.6.5/package.db | wc
Loaded package environment from /home/phadej/.ghc/x86_64-linux-8.6.5/environments/default
   4473    4473  372073
runghc GhcPkgScript.hs ghc-pkg-8.6.5 /cabal/store/ghc-8.6.5/package.db  13,34s user 1,22s system 100% cpu 14,481 total
wc  0,05s user 0,38s system 2% cpu 14,480 total

-}
{- AFTER THE CHANGE

% time runghc ./GhcPkgScript.hs ghc-pkg /cabal/store/ghc-8.6.5/package.db | wc 
Loaded package environment from /code/shared-haskell/cabal/.ghc.environment.x86_64-linux-8.6.5
   4477    4477  372405
runghc ./GhcPkgScript.hs ghc-pkg /cabal/store/ghc-8.6.5/package.db  7,48s user 0,94s system 100% cpu 8,369 total
wc  0,05s user 0,38s system 5% cpu 8,369 total

-}
module Main (main) where

import Distribution.InstalledPackageInfo (InstalledPackageInfo (..))
import Distribution.Pretty (prettyShow)
import Distribution.Simple.Compiler (PackageDB (..))
import Distribution.Simple.Program.HcPkg (dump, HcPkgInfo (..))
import Distribution.Simple.Program.Types (ConfiguredProgram, simpleConfiguredProgram, ProgramLocation (..))
import Distribution.Verbosity (silent)
import System.Environment (getArgs)

main :: IO ()
main = do
    (ghcpkg : arg : _) <- getArgs

    let cp :: ConfiguredProgram
        cp = simpleConfiguredProgram "ghc-pkg" (FoundOnSystem ghcpkg)

    let hcPkgInfo = HcPkgInfo cp
           -- this boolean flags are irrelevant for dump
           False False False False False False False False
                       -- ^ -- package-conf toggler for very old GHCs
    ipis <- dump hcPkgInfo silent (SpecificPackageDB arg)
    mapM_ (putStrLn . prettyShow . installedUnitId) ipis
```